Boat Stun Mechanic and River Speed Modification
Updated Physics Layer detection.

Applied animation curves to the River Manager when slowing or speeding up the river speed and now the speed of the river follows that curve when increasing or decreasing in speed.

Added 'isHit' a check to the River Obstacle script and a 'IsCollected' check to the River Collectible.

The player boat now slows the River speed upon being hit by an obstacle.

Renamed 'PlayerAndBoatTriggersTestScript' to 'DevDebugUI' for future UI testing in the Player Controls test scene.

Future fine tuning will be required as the project expands.