Pause Game Functionality and Initial UI and Menu Manager
Implemented Pause Game functionality, allowing the player to pause the game by pressing a single key. Doing so will pause the game.
On paused, the pause screen will be revealed, allowing the player to navigate the selected buttons

To create the pause screen functionality, the Main Menu Manager was reworked into a Menu Manager, allowing both the games Main Menu and Pause Menu screens to be navigated by the player and implemented by designers (aka me).